move isFieldDefault to src/form.js
