+++ src/form.js
@@ -162,3 +162,28 @@ export function isDisabled(element) {
   }
   return false;
 }
+
+/**
+ * Checks if a form field is in its default state.
+ * @param {!Element} field
+ * @return {boolean}
+ */
+export function isFieldDefault(field) {
+  switch (field.type) {
+    case 'select-multiple':
+    case 'select-one':
+      const {options} = field;
+      for (let j = 0; j < options.length; j++) {
+        if (options[j].selected !== options[j].defaultSelected) {
+          return false;
+        }
+      }
+      break;
+    case 'checkbox':
+    case 'radio':
+      return field.checked === field.defaultChecked;
+    default:
+      return field.value === field.defaultValue;
+  }
+  return true;
+}
